aula8_updates

+++ src/Aula8/EX_1/VeiculoEletrico.java
@@ -2,5 +2,5 @@ package Aula8.EX_1;
 
 public interface VeiculoEletrico {
     double autonomia(); // devolve autonomia restante
-    void carregar(int percentagem); // simula um carregamento até ‘percentagem’ 
+    void carregar(Double percentagem); // simula um carregamento até ‘percentagem’ 
 }
